requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
onnxruntime>=1.16.0
